bot/requirements.txt
httpx>=0.27.0
cryptography>=42.0.0
python-dotenv>=1.0.0
supabase>=2.4.0
